finish migrating to createSelector
